requirements.txt
timm==0.6.7
data
einops==0.6.1
ftfy==6.1.1
iopath==0.1.10
numpy==1.24.3
peft==0.3.0
Pillow==9.5.0
PyYAML==6.0
regex==2022.10.31
torchvision==0.14.1
torchaudio==0.13.1
pytorchvideo
fvcore
decord==0.6.0
tqdm==4.64.1
transformers==4.29.1
bigmodelvis
gradio
mdtex2html
ipdb
plyfile
trimesh
sentencepiece
openxlab